Listen for GUI close
This serves both to allow plugins to more easily detect when their GUIs
are closed, and also to ensure that click listeners are unregistered and
slot actions are removed to prevent memory leaks.

Unless I'm mistaken, there's still a listener registered for each GUI
opened, as well as a bunch of slot actions stored in memory.  While
these should now be cleared up when the player sends a close inventory
packet, theoretically a malicious player could use a modded client to
not send close packets and spam open a bunch of GUIs.  This would
register a bunch of listeners and store a bunch of click actions and
GUIs in memory, potentially crashing the server.

That is purely theoretical though, I don't know if that would actually work.